Switch to size_t, ptrdiff_t for indexes, shapes, and strides.
In addition to allowing 64-bit sizes, this breaks the correspondence
between the size type and the difference type, which brings us in
line with the STL.  Unfortunately, we now have the same problems
the STL has in terms of having to worry about comparisons and
conversions between sizes and differences.

Huge thanks to Christoph Lassner for earlier versions of this
change, which made it much easier to find all the code that
needed to be changed.