Revert "Fix pybind11_test.py"
This reverts commit f46c0f0ff876ceab5aaa3286e5f6e86902e72feb.

This was a somewhat unportable fix for pybind11 2.1.x that
is probably only desired by LSST; we'll just make that a
patch on the LSST build, and encourage everyone else to
use pybind11 2.2.x.  LSST will of course switch to a newer
pybind11 eventually, too, but isn't quite ready yet.